+++ despasito/fit_parameters/data_classes/solubility_parameter.py
@@ -17,7 +17,7 @@ from despasito.fit_parameters.interface import ExpDataTemplate
 class Data(ExpDataTemplate):
 
     r"""
-    Object for saturation data. This data is evaluated with "sat_props". 
+    Object for hildebrand solubility parameters. This data is evaluated with "solubility_parameter". 
 
     Parameters
     ----------
@@ -27,20 +27,23 @@ class Data(ExpDataTemplate):
         * name : str, data type, in this case SatProps
         * calctype : str, Optional, default: 'sat_props
         * T : list, List of temperature values for calculation
-        * xi : list, List of liquid mole fractions used in saturation properties calculations, should be 1 for the molecule of focus and 0 for the rest.
+        * P : list, List of pressure values used in calculations
+        * xi : list, List of liquid mole fractions used in calculations.
         * weights : dict, A dictionary where each key is the header used in the exp. data file. The value associated with a header can be a list as long as the number of data points to multiply by the objective value associated with each point, or a float to multiply the objective value of this data set.
         * rhodict : dict, Optional, default: {"minrhofrac":(1.0 / 60000.0), "rhoinc":10.0, "vspacemax":1.0E-4}, Dictionary of options used in calculating pressure vs. mole fraction curves.
 
     Attributes
     ----------
     name : str
-        Data type, in this case SatProps
-    calctype : str, Optional, default: 'sat_props'
+        Data type, in this case SolubilityParam
+    calctype : str, Optional, default: 'solubility_parameter'
         Thermodynamic calculation type
     T : list
         List of temperature values for calculation
     xi : list
         List of liquid mole fractions, only one should be equal to 1.
+    P  : list
+        List of pressure.
     """
 
     def __init__(self, data_dict):
@@ -54,8 +57,8 @@ class Data(ExpDataTemplate):
             self._thermodict["calculation_type"] = data_dict["calctype"]
             self.calctype = data_dict["calctype"]
         else:
-            self.calctype = "sat_props"
-            self._thermodict["calculation_type"] = "sat_props"
+            self.calctype = "solubility_parameter"
+            self._thermodict["calculation_type"] = "solubility_parameter"
 
         try:
             self.weights = data_dict["weights"]
@@ -75,43 +78,30 @@ class Data(ExpDataTemplate):
             if key in self.weights:
                 if type(self.weights[key]) != float and len(self.weights[key]) != len(self._thermodict[key]):
                     raise ValueError("Array of weights for '{}' values not equal to number of experimental values given.".format(key))
-        if "rhov" in data_dict:
-            key = "rhov"
-            self._thermodict["rhov"] = data_dict["rhov"]
+        if "P" in data_dict:
+            self._thermodict["Plist"] = data_dict["P"]
+            if 'P' in self.weights:
+                self.weights['Plist'] = self.weights.pop('P')
+        if "delta" in data_dict:
+            key = 'delta'
+            self._thermodict["delta"] = data_dict["delta"]
             if key in self.weights:
                 if type(self.weights[key]) != float and len(self.weights[key]) != len(self._thermodict[key]):
                     raise ValueError("Array of weights for '{}' values not equal to number of experimental values given.".format(key))
-        if "P" in data_dict:
-            self._thermodict["Psat"] = data_dict["P"]
-            if 'P' in self.weights:
-                self.weights['Psat'] = self.weights.pop('P')
-        if "Psat" in data_dict:
-            self._thermodict["Psat"] = data_dict["Psat"]
 
-        key = "Psat"
-        if key in self.weights:
-            if type(self.weights[key]) != float and len(self.weights[key]) != len(self._thermodict[key]):
-                raise ValueError("Array of weights for '{}' values not equal to number of experimental values given.".format(key))
-
-        tmp = ["Tlist"]
+        tmp = ["Tlist", "delta"]
         if not all([x in self._thermodict.keys() for x in tmp]):
-            raise ImportError("Given saturation data, value(s) for T should have been provided.")
-
-        tmp = ["Psat","rhol","rhov"]
-        if not any([x in self._thermodict.keys() for x in tmp]):
-            raise ImportError("Given saturation data, values for Psat, rhol, and/or rhov should have been provided.")
+            raise ImportError("Given solubility data, value(s) for T and delta should have been provided.")
 
         for key in self._thermodict.keys():
             if key not in self.weights:
                 if key != 'calculation_type':
                     self.weights[key] = 1.0
 
-        logger.info("Data type 'sat_props' initiated with calctype, {}, and data types: {}.\nWeight data by: {}".format(self.calctype,", ".join(self._thermodict.keys()),self.weights))
+        logger.info("Data type 'solubility parameter' initiated with calctype, {}, and data types: {}.\nWeight data by: {}".format(self.calctype,", ".join(self._thermodict.keys()),self.weights))
 
         if 'rhodict' in data_dict:
             self._thermodict["rhodict"] = data_dict["rhodict"]
-        else:
-            self._thermodict["rhodict"] = {"minrhofrac":(1.0 / 80000.0), "rhoinc":10.0, "vspacemax":1.0E-4}
 
     def _thermo_wrapper(self, eos):
 
@@ -135,13 +125,13 @@ class Data(ExpDataTemplate):
                 raise ValueError("Ambiguous instructions. Include xi to define intended component to obtain saturation properties")
             else:
                 self._thermodict['xilist'] = np.array([[1.0] for x in range(len(self._thermodict['Tlist']))])
- 
+
         # Run thermo calculations
         try:
             output_dict = thermo(eos, self._thermodict)
-            output = [output_dict["Psat"],output_dict["rhol"],output_dict["rhov"]]
+            output = [output_dict["delta"],output_dict["rhol"]]
         except:
-            raise ValueError("Calculation of calc_Psat failed")
+            raise ValueError("Calculation of solubility_parameter failed")
 
         return output
 
@@ -171,15 +161,13 @@ class Data(ExpDataTemplate):
         phase_list = np.transpose(np.array(phase_list))
 
         # objective function
-        obj_value = np.zeros(3)
-        if "Psat" in self._thermodict:
-            obj_value[0] = np.sum((((phase_list[0] - self._thermodict['Psat']) / self._thermodict['Psat'])**2)*self.weights['Psat'])
+        obj_value = np.zeros(2)
+        if "delta" in self._thermodict:
+            obj_value[0] = np.sum((((phase_list[0] - self._thermodict["delta"]) / self._thermodict["delta"])**2)*self.weights['delta'])
         if "rhol" in self._thermodict:
-            obj_value[1] = np.sum((((phase_list[1] - self._thermodict['rhol']) / self._thermodict['rhol'])**2)*self.weights['rhol'])
-        if "rhov" in self._thermodict:
-            obj_value[2] = np.sum((((phase_list[2] - self._thermodict['rhov']) / self._thermodict['rhov'])**2)*self.weights['rhov'])
+            obj_value[1] = np.sum((((phase_list[1] - self._thermodict["rhol"]) / self._thermodict["rhol"])**2)*self.weights['rhol'])
 
-        logger.debug("Obj. breakdown for {}: Psat {}, rhol {}, rhov {}".format(self.name,obj_value[0],obj_value[1],obj_value[2]))
+        logger.debug("Obj. breakdown for {}: delta {}, rhol {}".format(self.name,obj_value[0],obj_value[1]))
 
         if all(np.isnan(obj_value)):
             obj_total = np.nan